requirements.txt
discord.py~=1.5.0
Google-Images-Search~=1.3.3
jsonschema~=3.2.0
markovify~=0.8.3
numpy~=1.19.2
Pillow~=7.1.2
requests~=2.24.0
validator-collection~=1.4.2
windows-curses~=2.1.0
pygifsicle~=1.0.2
